Move annotations to another package
